bug with removing directory fixed

+++ Spark.py
@@ -1,6 +1,7 @@
 import json
 import os
 import glob
+import shutil
 import warnings
 
 from pyspark.ml.feature import Imputer, VectorAssembler
@@ -17,11 +18,17 @@ tempFilePath = os.path.join(projectPath, "input files/tempFile.csv")
 
 
 class Spark:
-    def __init__(self, csv):
+    def __init__(self, csv, fileName):
         self.csv = csv
+        self.fileName = fileName.split('.')[0] + '(modified).' + fileName.split('.')[-1]
+        print('fine name is : {}'.format(self.fileName))
         f = open(tempFilePath, "a")
         f.write(csv)
         f.close()
+
+    def deleteIfDirExists(self, dirPath):
+        if os.path.exists(dirPath) and os.path.isdir(dirPath):
+            shutil.rmtree(dirPath)
 
     def modifyDict(self, **kwargs):
         spark = SparkSession.builder.appName('Practice').getOrCreate()
@@ -30,12 +37,10 @@ class Spark:
             dfPyspark = dfPyspark.na.drop(how='any')
         if kwargs['min_avgSpeed'] is not None:
             dfPyspark = dfPyspark.filter("avgSpeed>=" + str(kwargs['min_avgSpeed']))
-        # dfPyspark.write.option("header", True).csv(os.path.join(outputFilesPath, inputFile.split('/')[-1]))
-        # print(dfPyspark.show(5))
+        self.deleteIfDirExists(os.path.join(outputFilesPath, self.fileName))
+        dfPyspark.write.option("header", True).csv(os.path.join(outputFilesPath, self.fileName))
         return dfPyspark
 
     def removeTempFile(self):
-        try:
+        if os.path.exists(tempFilePath):
             os.remove(tempFilePath)
-        except:
-            pass
